Add fix for in-source build of cppkafka.h.
This fixes an issue where a cppkafka.h is being generated directly
within the source directory, which causes issues on sandboxed build
environments where the source directory is mounted readonly.

This PR changes the configure_file() directive to point to the
binary directory, and uses the install directives to move the
generated file into the build output.